graph-system/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
networkx==3.3
google-generativeai==0.8.3
python-dotenv==1.0.1
openpyxl==3.1.5
python-multipart==0.0.9
pydantic==2.9.2
